made initial changes to make pool class resemble environment class from spike exchange miniapp

+++ neuromapp/coreneuron_1.0/queueing/queue.cpp
@@ -33,7 +33,7 @@
 
 namespace queueing {
 
-void queue::insert(double tt, double d) {
+void queue::insert(double tt, int d) {
     pq_que.push(event(d,tt));
 }
